+++ pubtools/pulplib/_impl/client/poller.py
@@ -91,6 +91,8 @@ class TaskPoller(object):
                 return True
 
             if task.completed and not task.succeeded:
+                LOG.warning("Pulp task failed: %s", task.id)
+
                 exception = TaskFailedException(task)
                 descriptor.yield_exception(exception)
                 return True
@@ -101,6 +103,8 @@ class TaskPoller(object):
             if not task.completed:
                 # can't resolve the future yet since there's a pending task
                 return False
+
+            LOG.info("Pulp task completed: %s", task.id)
 
         # OK, future can be resolved with the completed tasks
         descriptor.yield_result(out)
@@ -153,6 +157,18 @@ class TaskPoller(object):
                 descriptor.yield_exception(ex)
 
         return descriptor_tasks, all_tasks
+
+    def cancel(self, task_data):
+        task_ids = [t["task_id"] for t in task_data["spawned_tasks"]]
+
+        for task_id in task_ids:
+            url = os.path.join(self.url, "pulp/api/v2/tasks/%s/" % task_id)
+            response = self.session.delete(url)
+            response.raise_for_status()
+
+            LOG.info("Cancelled Pulp task: %s", task_id)
+
+        return True
 
     def log_if_inactive(self):
         now = self.timer()